pdf2xl-backend/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
cryptography==42.0.2
cssselect==1.2.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.1
Flask-Cors==4.0.0
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.3
pdfquery==0.4.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pypdfium2==4.26.0
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.4
pyzmq==25.1.2
requests==2.31.0
roman==4.1
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
XlsxWriter==3.1.9
